Split Rumba library into focused modules
